requirements.txt
streamlit==1.50.0
pandas==2.3.2
reportlab==4.4.4
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.0